passes: move PassPipelineParseError out to exceptions.py
