+++ xdsl/utils/exceptions.py
@@ -10,6 +10,8 @@ from dataclasses import dataclass
 from io import StringIO
 from typing import Any, IO
 
+from utils.parse_pipeline import Token
+
 if typing.TYPE_CHECKING:
     from xdsl.parser import Span, BacktrackingHistory
     from xdsl.ir import Attribute
@@ -158,3 +160,11 @@ class MultipleSpansParseError(ParseError):
         print(self.ref_text or "With respect to:", file=file)
         for span, msg in self.refs:
             print(span.print_with_context(msg), file=file)
+
+
+class PassPipelineParseError(BaseException):
+    def __init__(self, token: Token, msg: str):
+        super().__init__(
+            "Error parsing pass pipeline specification:\n"
+            + token.span.print_with_context(msg)
+        )
